mvlc_mvp: implement possible workaround for some MDPP32 modules.
Certain combinations of flash chip and flash firmware sometimes do not
return the first request word in responses. When checking responses a new
workaround is to pop off the first word of the request and try the response
check again.

Untested as we do not have access to the exact module causing the problem!